+++ adapters/eclipse.pl
@@ -4,7 +4,7 @@
 %  Copyright (c) 1998-2013 Paulo Moura <pmoura@logtalk.org>
 %
 %  Adapter file for ECLiPSe 6.1#143 and later versions
-%  Last updated on May 9, 2013
+%  Last updated on May 17, 2013
 %
 %  This program is free software: you can redistribute it and/or modify
 %  it under the terms of the GNU General Public License as published by
@@ -172,6 +172,12 @@ setup_call_cleanup(_, _, _) :-
 % '$lgt_prolog_meta_directive'(@callable, @callable)
 
 '$lgt_prolog_meta_directive'(_, _) :-
+	fail.
+
+
+% '$lgt_prolog_to_logtalk_meta_argument_specifier_hook'(@nonvar, -atom)
+
+'$lgt_prolog_to_logtalk_meta_argument_specifier_hook'(_, _) :-
 	fail.